add distributed test for grow function
had a discrepancy between the encrypted and test data for performance
in lattices of indices: ( 10 - 20, 10 - 20). This shouldn't have
been showing up I thought so i tested  the encrypt and verified it was
correct. it was, then i tested the vh-association groups of some indices
between both sets. They were the same, so the map function should
have been the same. Turns out the encrypted anchors were just randomly
distributed and the test data were centered. So because the lattices
can grow more in the center ( more room to grow ) the encrypted datas
average growth was smaller, with a lower percentage of groups over size (2,2)

But the lifecycle function works great.
next: build a lattice transform function that takes the lattice and
converts it to a transform function to be held against a picture.