add uncertain growth
for the mnist dataset the scarcity and unreliability of the information was
too little for the standard build and bind functions. So i made functions
that work with less reliable and more scarce information- as well as
a check that will determine if an inputted dataset should be classified as
uncertain or not. I check that by cycling through the vh groups
and if there is a neighbour with perfect association in the fourth slot then
that shows there are atleast 4 indices with the exact same values as the
standard index.
These are usually unused nodes, so in the final file i will make a function
that uses these as a buffer.

There are also a bunch of animations demonstrating the function, not the uncertain
version though. I will try to add those soon. I also ran some more correlation
and gather testing. these were related to investigating why the mnist testing
went so poorly and I was trying unsuccessfully to download a more detailed dataset.